+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+-   Properly handle when 'keep log count' setting is zero.
+
 ## [0.14.2] - 2025-09-01
 
 ### Fixed